Removing loadByFamilyId() from OAuth2TokenCache - moving to specific FOCI implementation subclass
