+++ common/src/androidTest/java/com/microsoft/identity/common/MsalOAuth2TokenCacheTest.java
@@ -61,7 +61,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import static com.microsoft.identity.common.SharedPreferencesAccountCredentialCacheTest.AUTHORITY_TYPE;
 import static com.microsoft.identity.common.SharedPreferencesAccountCredentialCacheTest.CACHED_AT;
 import static com.microsoft.identity.common.SharedPreferencesAccountCredentialCacheTest.CLIENT_ID;
 import static com.microsoft.identity.common.SharedPreferencesAccountCredentialCacheTest.ENVIRONMENT;
@@ -107,8 +106,8 @@ public class MsalOAuth2TokenCacheTest extends AndroidSecretKeyEnabledHelper {
             MicrosoftAccount,
             MicrosoftRefreshToken> mockCredentialAdapter;
 
-    private AccountCredentialTestBundle defaultTestBundle;
-    private IAccountCredentialCache accountCredentialCache;
+    AccountCredentialTestBundle defaultTestBundle;
+    IAccountCredentialCache accountCredentialCache;
 
     public static class AccountCredentialTestBundle {
 
@@ -831,291 +830,6 @@ public class MsalOAuth2TokenCacheTest extends AndroidSecretKeyEnabledHelper {
         assertEquals(defaultTestBundle.mGeneratedAccessToken, secondaryLoad.getAccessToken());
         assertEquals(defaultTestBundle.mGeneratedRefreshToken, secondaryLoad.getRefreshToken());
         assertNull(secondaryLoad.getIdToken());
-    }
-
-    @Test
-    public void testRetrieveFrt() throws ClientException {
-        final String randomHomeAccountId = UUID.randomUUID().toString();
-
-        final AccountCredentialTestBundle frtTestBundle = new AccountCredentialTestBundle(
-                MicrosoftAccount.AUTHORITY_TYPE_V1_V2,
-                UUID.randomUUID().toString(),
-                "test.user@tenant.onmicrosoft.com",
-                randomHomeAccountId,
-                ENVIRONMENT,
-                UUID.randomUUID().toString(),
-                TARGET,
-                CACHED_AT,
-                EXPIRES_ON,
-                SECRET,
-                CLIENT_ID,
-                SECRET,
-                MicrosoftStsAccountCredentialAdapterTest.MOCK_ID_TOKEN_WITH_CLAIMS,
-                "1"
-        );
-
-        when(
-                mockCredentialAdapter.createAccount(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedAccount);
-
-        when(
-                mockCredentialAdapter.createAccessToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedAccessToken);
-
-        when(
-                mockCredentialAdapter.createRefreshToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedRefreshToken);
-
-        when(
-                mockCredentialAdapter.createIdToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedIdToken);
-
-        // Save the family token data
-        mOauth2TokenCache.save(
-                mockStrategy,
-                mockRequest,
-                mockResponse
-        );
-
-        final ICacheRecord familyCacheRecord = mOauth2TokenCache.loadByFamilyId(
-                null,
-                null,
-                frtTestBundle.mGeneratedAccount
-        );
-
-        assertNotNull(familyCacheRecord);
-        assertNotNull(familyCacheRecord.getAccount());
-        assertNotNull(familyCacheRecord.getRefreshToken());
-        assertNull(familyCacheRecord.getIdToken());
-        assertNull(familyCacheRecord.getAccessToken());
-
-        final ICacheRecord familyCacheRecordWithClientId = mOauth2TokenCache.loadByFamilyId(
-                CLIENT_ID,
-                null,
-                frtTestBundle.mGeneratedAccount
-        );
-
-        assertNotNull(familyCacheRecordWithClientId);
-        assertNotNull(familyCacheRecordWithClientId.getAccount());
-        assertNotNull(familyCacheRecordWithClientId.getRefreshToken());
-        assertNotNull(familyCacheRecordWithClientId.getIdToken());
-        assertNotNull(familyCacheRecordWithClientId.getAccessToken());
-
-        final ICacheRecord familyCacheRecordWithClientIdButNonMatchingTarget =
-                mOauth2TokenCache.loadByFamilyId(
-                        CLIENT_ID,
-                        "foo",
-                        frtTestBundle.mGeneratedAccount
-                );
-
-        assertNotNull(familyCacheRecordWithClientIdButNonMatchingTarget);
-        assertNotNull(familyCacheRecordWithClientIdButNonMatchingTarget.getAccount());
-        assertNotNull(familyCacheRecordWithClientIdButNonMatchingTarget.getRefreshToken());
-        assertNotNull(familyCacheRecordWithClientIdButNonMatchingTarget.getIdToken());
-        assertNull(familyCacheRecordWithClientIdButNonMatchingTarget.getAccessToken());
-
-        final ICacheRecord wrongClientIdResult =
-                mOauth2TokenCache.loadByFamilyId(
-                        "12345",
-                        "foo",
-                        frtTestBundle.mGeneratedAccount
-                );
-
-        assertNotNull(wrongClientIdResult);
-        assertNotNull(wrongClientIdResult.getAccount());
-        assertNotNull(wrongClientIdResult.getRefreshToken());
-        assertNull(wrongClientIdResult.getIdToken());
-        assertNull(wrongClientIdResult.getAccessToken());
-    }
-
-    @Test
-    public void testOnlyOneFrtMayExistAcrossClientsForAccount() throws ClientException {
-        // Save an FRT
-        final String randomHomeAccountId = UUID.randomUUID().toString();
-        final String localAccountId = UUID.randomUUID().toString();
-        final String realm = UUID.randomUUID().toString();
-
-        final AccountCredentialTestBundle frtTestBundle = new AccountCredentialTestBundle(
-                MicrosoftAccount.AUTHORITY_TYPE_V1_V2,
-                localAccountId,
-                "test.user@tenant.onmicrosoft.com",
-                randomHomeAccountId,
-                ENVIRONMENT,
-                realm,
-                TARGET,
-                CACHED_AT,
-                EXPIRES_ON,
-                SECRET,
-                CLIENT_ID,
-                SECRET,
-                MicrosoftStsAccountCredentialAdapterTest.MOCK_ID_TOKEN_WITH_CLAIMS,
-                "1"
-        );
-
-        when(
-                mockCredentialAdapter.createAccount(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedAccount);
-
-        when(
-                mockCredentialAdapter.createAccessToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedAccessToken);
-
-        when(
-                mockCredentialAdapter.createRefreshToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedRefreshToken);
-
-        when(
-                mockCredentialAdapter.createIdToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle.mGeneratedIdToken);
-
-        mOauth2TokenCache.save(
-                mockStrategy,
-                mockRequest,
-                mockResponse
-        );
-
-        // Save another FRT, this time with a different client id
-        final AccountCredentialTestBundle frtTestBundle2 = new AccountCredentialTestBundle(
-                MicrosoftAccount.AUTHORITY_TYPE_V1_V2,
-                localAccountId,
-                "test.user@tenant.onmicrosoft.com",
-                randomHomeAccountId,
-                ENVIRONMENT,
-                realm,
-                TARGET,
-                CACHED_AT,
-                EXPIRES_ON,
-                SECRET,
-                CLIENT_ID + "2",
-                SECRET,
-                MicrosoftStsAccountCredentialAdapterTest.MOCK_ID_TOKEN_WITH_CLAIMS,
-                "1"
-        );
-
-        when(
-                mockCredentialAdapter.createAccount(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle2.mGeneratedAccount);
-
-        when(
-                mockCredentialAdapter.createAccessToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle2.mGeneratedAccessToken);
-
-        when(
-                mockCredentialAdapter.createRefreshToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle2.mGeneratedRefreshToken);
-
-        when(
-                mockCredentialAdapter.createIdToken(
-                        mockStrategy,
-                        mockRequest,
-                        mockResponse
-                )
-        ).thenReturn(frtTestBundle2.mGeneratedIdToken);
-
-        // Save the family token data
-        mOauth2TokenCache.save(
-                mockStrategy,
-                mockRequest,
-                mockResponse
-        );
-
-        // Test only one FRT exists and it is the second one saved...
-        final ICacheRecord cacheRecord = mOauth2TokenCache.loadByFamilyId(
-                CLIENT_ID,
-                null,
-                frtTestBundle2.mGeneratedAccount
-        );
-
-        assertNotNull(cacheRecord);
-        assertNotNull(cacheRecord.getRefreshToken());
-        assertNotNull(cacheRecord.getAccessToken());
-        assertNotNull(cacheRecord.getIdToken());
-        assertEquals(
-                CLIENT_ID + "2",
-                cacheRecord.getRefreshToken().getClientId()
-        );
-
-        // Check querying for the FRT in the second app yields the same FRT
-        final ICacheRecord cacheRecord2 = mOauth2TokenCache.loadByFamilyId(
-                CLIENT_ID + "2",
-                null,
-                frtTestBundle2.mGeneratedAccount
-        );
-
-        assertNotNull(cacheRecord2);
-        assertNotNull(cacheRecord2.getRefreshToken());
-        assertNotNull(cacheRecord2.getAccessToken());
-        assertNotNull(cacheRecord2.getIdToken());
-        assertEquals(
-                CLIENT_ID + "2",
-                cacheRecord2.getRefreshToken().getClientId()
-        );
-
-        // Test querying with a different account yields nothing at all....
-
-        final AccountRecord randomAcct = new AccountRecord();
-        randomAcct.setAuthorityType(AUTHORITY_TYPE);
-        randomAcct.setLocalAccountId(UUID.randomUUID().toString());
-        randomAcct.setUsername("foo@bar.com");
-        randomAcct.setHomeAccountId(UUID.randomUUID().toString());
-        randomAcct.setEnvironment(ENVIRONMENT);
-        randomAcct.setRealm(REALM);
-
-        final ICacheRecord cacheRecord3 = mOauth2TokenCache.loadByFamilyId(
-                CLIENT_ID + "2",
-                null,
-                randomAcct
-        );
-
-        assertNotNull(cacheRecord3);
-        assertNotNull(cacheRecord3.getAccount());
-        assertNull(cacheRecord3.getRefreshToken());
-        assertNull(cacheRecord3.getAccessToken());
-        assertNull(cacheRecord3.getIdToken());
     }
 
 }
