+++ common/src/main/java/com/microsoft/identity/common/internal/cache/ADALOAuth2TokenCache.java
@@ -176,15 +176,6 @@ public class ADALOAuth2TokenCache
     }
 
     @Override
-    public ICacheRecord loadByFamilyId(final String clientId,
-                                       final String target,
-                                       final AccountRecord accountRecord) {
-        throw new UnsupportedOperationException(
-                ERR_UNSUPPORTED_OPERATION
-        );
-    }
-
-    @Override
     public ICacheRecord load(
             final String clientId,
             final String target,
