+++ common/src/main/java/com/microsoft/identity/common/internal/providers/oauth2/OAuth2TokenCache.java
@@ -75,22 +75,6 @@ public abstract class OAuth2TokenCache
     );
 
     /**
-     * Loads the tokens available for the supplied client criteria.
-     *
-     * @param clientId      The current client's id.
-     * @param target        The desired scopes.
-     * @param accountRecord The current account.
-     * @return An ICacheRecord containing the account. If a matching id token is available
-     * (for the provided clientId), it is returned. If a matching access token is available
-     * (for the provided client id), it is also returned. If a matching refresh token is available
-     * it is returned.
-     */
-    public abstract ICacheRecord loadByFamilyId(final String clientId,
-                                                final String target,
-                                                final AccountRecord accountRecord
-    );
-
-    /**
      * Loads the tokens for the supplied Account into the result {@link ICacheRecord}.
      *
      * @param clientId The ClientId of the current app.
